Fixes required for ROS Melodic support
- OpenCV2 to 3 migration
- Retrain laser detector after OpenCV migration, as old models are incompatible
- Replace deprecated pluginlib macro
- Qt5 support for Rviz plugins
- Replace tf by tf2 message_filter in rviz plugin (contributed by user @f-fl0 in srl-freiburg/spencer_tracking_rviz_plugin)
- Add boost namespace for shared_ptr (for rviz plugin, this was contributed by @makokal in srl-freiburg/spencer_tracking_rviz_plugin)
- Update Dockerfiles

+++ tracking/people/srl_nearest_neighbor_tracker/src/srl_nearest_neighbor_tracker/ros/geometry_utils.cpp
@@ -249,7 +249,7 @@ bool GeometryUtils::posePassesSanityCheck(const geometry_msgs::PoseWithCovarianc
 
     // Position
     const float MAX_REASONABLE_DISTANCE_FROM_ORIGIN = 1000.0f; // in meters
-    float positionValues[] = { pose.position.x, pose.position.y, pose.position.z };
+    double positionValues[] = { pose.position.x, pose.position.y, pose.position.z };
     for(size_t i = 0; i < sizeof(positionValues) / sizeof(positionValues[0]); i++) {
         if(!isfinite(positionValues[i]) || abs(positionValues[i]) > MAX_REASONABLE_DISTANCE_FROM_ORIGIN) {
             ROS_WARN("Suspicious coordinate value(s) in pose position encountered!");
@@ -259,8 +259,8 @@ bool GeometryUtils::posePassesSanityCheck(const geometry_msgs::PoseWithCovarianc
 
     // Orientation
     if(checkOrientation) {
-        float squaredSum = 0;
-        float orientationValues[] = { pose.orientation.x, pose.orientation.y, pose.orientation.z, pose.orientation.w };
+        double squaredSum = 0;
+        double orientationValues[] = { pose.orientation.x, pose.orientation.y, pose.orientation.z, pose.orientation.w };
         for(size_t i = 0; i < sizeof(orientationValues) / sizeof(orientationValues[0]); i++) {
             if(!isfinite(orientationValues[i])) {
                 ROS_WARN("Non-finite pose orientation value(s) encountered!");
